feat: Refactor architecture with Gateway layer and Router system
Major architectural refactoring to introduce Gateway layer and Router system:

- Add Gateway abstraction layer (gateway/gateway.h, gateway_manager)
  - Support HTTP, WebSocket, and STDIO transports
  - Unified gateway interface for all channels
  - Platform layer abstraction for HTTP and WebSocket

- Add Router system (router/router.h, router.c)
  - Replaces Input Processor for better semantic clarity
  - Unified routing for commands and chat messages
  - Gateway-to-channel mapping support

- Migrate channels to use Gateway layer
  - CLI Channel: uses STDIO Gateway
  - Telegram Channel: uses HTTP Gateway
  - WebSocket Channel: uses WebSocket Server Gateway
  - Feishu Channel: uses WebSocket Client Gateway
  - QQ Channel: uses WebSocket Client Gateway

- Remove old input_processor (replaced by Router)
- Update architecture documentation
- Fix STDIO gateway initialization issue

This refactoring enables:
- Cleaner separation of transport and business logic
- Easier addition of new channels and protocols
- Platform-independent gateway implementations

+++ main/cli/editor.c
@@ -303,14 +303,14 @@ static void cli_do_completion(cli_terminal_t *term)
         /* Extract prefix after "/" */
         int prefix_start = 1;
         int prefix_len = term->cursor - prefix_start;
-        if (prefix_len > 0 && prefix_len < sizeof(prefix) - 1) {
+        if (prefix_len > 0 && prefix_len < (int)sizeof(prefix) - 1) {
             memcpy(prefix, &term->line[prefix_start], prefix_len);
             prefix[prefix_len] = '\0';
         }
     } else {
         /* Regular word completion */
         int word_len = term->cursor - word_start;
-        if (word_len > 0 && word_len < sizeof(prefix) - 1) {
+        if (word_len > 0 && word_len < (int)sizeof(prefix) - 1) {
             memcpy(prefix, &term->line[word_start], word_len);
             prefix[word_len] = '\0';
         }
